move merged_momentum_op to optimizer dir
